feat(croquis)!: expose parse-only SFC facade
