+++ crates/vize_atelier_sfc/src/lib.rs
@@ -44,13 +44,26 @@ pub mod compile_template;
 pub mod croquis;
 pub mod css;
 pub mod module_shape;
-pub mod parse;
 pub mod rewrite_default;
 pub mod script;
 pub mod source_map;
 pub mod style;
 pub mod types;
 pub mod vite_plugin;
+
+/// Backward-compatible SFC parser module.
+pub mod parse {
+    use crate::types::{SfcDescriptor, SfcError, SfcParseOptions};
+
+    /// Parse a Vue SFC into a zero-copy descriptor.
+    #[inline]
+    pub fn parse_sfc<'a>(
+        source: &'a str,
+        options: SfcParseOptions,
+    ) -> Result<SfcDescriptor<'a>, SfcError> {
+        vize_croquis::sfc::parse_sfc(source, options)
+    }
+}
 
 // Re-exports for public API
 pub use bundler::{
