+++ crates/vize_croquis/Cargo.toml
@@ -18,6 +18,7 @@ serde.workspace = true
 serde_json.workspace = true
 dashmap.workspace = true
 regex.workspace = true
+memchr.workspace = true
 
 # OXC for high-performance AST parsing
 oxc_parser.workspace = true
