+++ crates/vize_croquis/src/sfc/parse/block/compat.rs
