+++ crates/vize_croquis/src/sfc/parse/template_boundary/fast_path.rs
