+++ crates/vize_croquis/src/sfc/parse/tests.rs
